pyproject.toml
[project]
name = "whole-sys"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.16.5",
    "fastapi>=0.117.1",
    "passlib>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.43",
    "uvicorn>=0.37.0",
]
